Exceptions,Validations and Pagination to APIs for Product Entity

+++ src/main/java/com/ecommerce/project/springbootecom/service/ProductServiceImpl.java
@@ -1,5 +1,6 @@
 package com.ecommerce.project.springbootecom.service;
 
+import com.ecommerce.project.springbootecom.exception.APIException;
 import com.ecommerce.project.springbootecom.exception.ResourceNotFoundException;
 import com.ecommerce.project.springbootecom.model.Category;
 import com.ecommerce.project.springbootecom.model.Product;
@@ -11,6 +12,10 @@ import com.ecommerce.project.springbootecom.repo.ProductRepo;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -42,52 +47,112 @@ public class ProductServiceImpl implements ProductService{
 
         Category category=categoryRepo.findById(categoryId).orElseThrow(()->
                                                 new ResourceNotFoundException("Category","CategoryId",categoryId));
-        Product product=modelMapper.map(productDTO, Product.class);
-        product.setCategory(category);
-        product.setProductImage("default.png");
-        Double specialPrice=product.getProductPrice()-(product.getProductPrice()*product.getDiscount()*0.01);
-        product.setSpecialPrice(specialPrice);
+        boolean isProductNotPresent=true;
 
-        Product savedProduct=productRepo.save(product);
-        return modelMapper.map(savedProduct, ProductDTO.class);
+
+        for(Product product: category.getProducts()){
+            if(product.getProductName().equals(productDTO.getProductName())){
+                isProductNotPresent=false;
+                break;
+            }
+        }
+
+        if(isProductNotPresent){
+            Product product = modelMapper.map(productDTO, Product.class);
+            product.setCategory(category);
+            product.setProductImage("default.png");
+            Double specialPrice = product.getProductPrice() - (product.getProductPrice() * product.getDiscount() * 0.01);
+            product.setSpecialPrice(specialPrice);
+
+            Product savedProduct = productRepo.save(product);
+            return modelMapper.map(savedProduct, ProductDTO.class);
+        }
+        else{
+            throw new APIException("Product already exist in Database!!");
+        }
     }
 
     @Override
-    public ProductResponse getAllProducts() {
+    public ProductResponse getAllProducts(Integer pageNumber,Integer pageSize,String sortBy,String sortOrder) {
 
-        List<Product> allProducts=productRepo.findAll();
+        Sort sortByAndOrder = sortOrder.equalsIgnoreCase("asc")
+                                ?Sort.by(sortBy).ascending()
+                                :Sort.by(sortBy).descending();
+
+        Pageable pageDetails= PageRequest.of(pageNumber,pageSize,sortByAndOrder);
+        Page<Product> productList=productRepo.findAll(pageDetails);
+        List<Product> allProducts=productList.getContent();
+
+        if(allProducts.isEmpty()){
+            throw new APIException("No products found!");
+        }
         ProductResponse productResponse=new ProductResponse();
 
         List<ProductDTO> productDTOS=allProducts.stream()
                 .map(product -> modelMapper.map(product,ProductDTO.class)).toList();
         productResponse.setContent(productDTOS);
+        productResponse.setPageNumber(productList.getNumber());
+        productResponse.setPageSize(productList.getSize());
+        productResponse.setTotalPages(productList.getTotalPages());
+        productResponse.setTotalElements(productList.getTotalElements());
+        productResponse.setLastPage(productList.isLast());
+
         return productResponse;
     }
 
+
     @Override
-    public ProductResponse getProductsByCategory(Long categoryId) {
+    public ProductResponse getProductsByCategory(Long categoryId, Integer pageNumber, Integer pageSize, String sortBy, String sortOrder) {
         Category category=categoryRepo.findById(categoryId)
                         .orElseThrow(()->new ResourceNotFoundException("Category","CategoryId",categoryId));
-        List<Product> productListByCategory=productRepo.findByCategoryOrderByProductPriceAsc(category);
+        Sort sortByAndOrder=sortOrder.equalsIgnoreCase("asc")
+                            ?Sort.by(sortBy).ascending()
+                            :Sort.by(sortBy).descending();
+        Pageable pageDetails =PageRequest.of(pageNumber,pageSize,sortByAndOrder);
+        Page<Product> productList=productRepo.findByCategory(category,pageDetails);
+
+        List<Product> productListByCategory=productList.getContent();
+        if(productListByCategory.isEmpty()){
+            throw new APIException("No products found in Database with the given category!!");
+        }
 
         List<ProductDTO> allProductsListByCategoryDTOs= productListByCategory.stream()
                 .map(product -> modelMapper.map(product,ProductDTO.class)).toList();
 
         ProductResponse productResponse=new ProductResponse();
         productResponse.setContent(allProductsListByCategoryDTOs);
+        productResponse.setPageNumber(productList.getNumber());
+        productResponse.setPageSize(productList.getSize());
+        productResponse.setTotalPages(productList.getTotalPages());
+        productResponse.setTotalElements(productList.getTotalElements());
+        productResponse.setLastPage(productList.isLast());
         return productResponse;
     }
 
     @Override
-    public ProductResponse getProductsByKeyword(String keyword) {
+    public ProductResponse getProductsByKeyword(String keyword, Integer pageNumber, Integer pageSize, String sortBy, String sortOrder) {
         //%Keyword%
-        List<Product> productListByCategory=productRepo.findByProductNameLikeIgnoreCase('%'+keyword+'%');
+        Sort sortByAndOrder=sortOrder.equalsIgnoreCase("asc")
+                ?Sort.by(sortBy).ascending()
+                :Sort.by(sortBy).descending();
+        Pageable pageDetails =PageRequest.of(pageNumber,pageSize,sortByAndOrder);
+        Page<Product> productList=productRepo.findByProductNameLikeIgnoreCase('%'+keyword+'%',pageDetails);
+
+        List<Product> productListByCategory=productList.getContent();
+        if(productListByCategory.isEmpty()){
+            throw new APIException("No products found in Database with the given Keyword!!");
+        }
 
         List<ProductDTO> allProductsListByCategoryDTOs= productListByCategory.stream()
                 .map(product -> modelMapper.map(product,ProductDTO.class)).toList();
 
         ProductResponse productResponse=new ProductResponse();
         productResponse.setContent(allProductsListByCategoryDTOs);
+        productResponse.setPageNumber(productList.getNumber());
+        productResponse.setPageSize(productList.getSize());
+        productResponse.setTotalPages(productList.getTotalPages());
+        productResponse.setTotalElements(productList.getTotalElements());
+        productResponse.setLastPage(productList.isLast());
         return productResponse;
     }
 
@@ -110,6 +175,9 @@ public class ProductServiceImpl implements ProductService{
     @Override
     public ProductDTO deleteProduct(Long productId) {
         Product savedProduct=productRepo.findById(productId).orElseThrow(()->new ResourceNotFoundException("Product","productId",productId));
+        if(savedProduct==null){
+            throw new ResourceNotFoundException("Product","productId",productId);
+        }
         productRepo.delete(savedProduct);
         return modelMapper.map(savedProduct,ProductDTO.class);
     }
